MNT: merge and modernize 00, 01, 02
