+++ profile_collection/startup/00-startup.py
@@ -1,39 +1,23 @@
-import sys
-import logging
+# Make ophyd listen to pyepics.
+from ophyd import setup_ophyd
+setup_ophyd()
 
-import matplotlib.pyplot as plt
-import numpy as np
-import pandas as pd
-import faulthandler
-from metadatastore.mds import MDS
+# Set up a RunEngine and use metadata backed by a sqlite file.
+from bluesky import RunEngine
+from bluesky.utils import get_history
+RE = RunEngine(get_history())
+RE.md['group'] = ''
+RE.md['config'] = {}
+RE.md['beamline_id'] = 'HXN'
+RE.verbose = True
+
+
+# Set up a Broker.
+# TODO clean this up
 from databroker import Broker
+from databroker.headersource.mongo import MDS
+from databroker.assets.mongo import Registry
 from databroker.core import register_builtin_handlers
-from filestore.fs import FileStore
-from hxntools.handlers.xspress3 import Xspress3HDF5Handler
-from hxntools.handlers.timepix import TimepixHDF5Handler
-
-faulthandler.enable()
-plt.ion()
-
-handler = logging.StreamHandler(sys.stderr)
-fmt = logging.Formatter("%(asctime)-15s [%(name)5s:%(levelname)s] %(message)s")
-handler.setFormatter(fmt)
-handler.setLevel(logging.INFO)
-
-logging.getLogger('hxntools').addHandler(handler)
-logging.getLogger('hxnfly').addHandler(handler)
-logging.getLogger('ppmac').addHandler(handler)
-
-logging.getLogger('hxnfly').setLevel(logging.DEBUG)
-logging.getLogger('hxntools').setLevel(logging.DEBUG)
-logging.getLogger('ppmac').setLevel(logging.INFO)
-
-
-# Flyscan results are shown using pandas. Maximum rows/columns to use when
-# printing the table:
-pd.options.display.width = 180
-pd.options.display.max_rows = None
-pd.options.display.max_columns = 10
 
 _mds_config = {'host': 'xf03id-ca1',
                'port': 27017,
@@ -44,7 +28,7 @@ mds = MDS(_mds_config, auth=False)
 _fs_config = {'host': 'xf03id-ca1',
               'port': 27017,
               'database': 'filestore-new'}
-db_new = Broker(mds, FileStore(_fs_config))
+db_new = Broker(mds, Registry(_fs_config))
 
 _mds_config_old = {'host': 'xf03id-ca1',
                    'port': 27017,
@@ -55,21 +39,11 @@ mds_old = MDS(_mds_config_old, auth=False)
 _fs_config_old = {'host': 'xf03id-ca1',
                   'port': 27017,
                   'database': 'filestore'}
-db_old = Broker(mds_old, FileStore(_fs_config_old))
+db_old = Broker(mds_old, Registry(_fs_config_old))
 
+RE.subscribe(db_new.insert)
 
-def _hxn_register_handlers(inp_db):
-    "helper function to register handlers to both assert registries"
-    register_builtin_handlers(inp_db.fs)
-    inp_db.fs.register_handler(Xspress3HDF5Handler.HANDLER_NAME,
-                               Xspress3HDF5Handler)
-    inp_db.fs.register_handler(TimepixHDF5Handler._handler_name,
-                               TimepixHDF5Handler, overwrite=True)
-
-
-_hxn_register_handlers(db_new)
-_hxn_register_handlers(db_old)
-del _hxn_register_handlers
+import pandas as pd
 
 
 # wrapper for two databases
@@ -95,7 +69,88 @@ class Broker_New(Broker):
         return result
 
 
-db = Broker_New(mds, FileStore(_fs_config))
+db = Broker_New(mds, Registry(_fs_config))
+
+from hxntools.handlers.xspress3 import Xspress3HDF5Handler
+from hxntools.handlers.timepix import TimepixHDF5Handler
+
+
+def _hxn_register_handlers(inp_db):
+    "helper function to register handlers to both assert registries"
+    register_builtin_handlers(inp_db.fs)
+    inp_db.fs.register_handler(Xspress3HDF5Handler.HANDLER_NAME,
+                               Xspress3HDF5Handler)
+    inp_db.fs.register_handler(TimepixHDF5Handler._handler_name,
+                               TimepixHDF5Handler, overwrite=True)
+
+
+_hxn_register_handlers(db_new)
+_hxn_register_handlers(db_old)
+del _hxn_register_handlers
+# do the rest of the standard configuration
+from IPython import get_ipython
+from nslsii import configure_base, configure_olog
+
+configure_base(get_ipython().user_ns, db_new)
+configure_olog(get_ipython().user_ns)
+
+
+# set up some HXN specific callbacks
+from ophyd.callbacks import UidPublish
+from hxntools.scan_number import HxnScanNumberPrinter
+from hxntools.scan_status import HxnScanStatus
+from ophyd import EpicsSignal
+
+uid_signal = EpicsSignal('XF:03IDC-ES{BS-Scan}UID-I', name='uid_signal')
+uid_broadcaster = UidPublish(uid_signal)
+scan_number_printer = HxnScanNumberPrinter()
+hxn_scan_status = HxnScanStatus('XF:03IDC-ES{Status}ScanRunning-I',
+                                name='hxn_scan_status')
+
+# Pass on only start/stop documents to a few subscriptions
+for _event in ('start', 'stop'):
+    RE.subscribe(scan_number_printer, _event)
+    RE.subscribe(uid_broadcaster, _event)
+    RE.subscribe(hxn_scan_status, _event)
+
+
+def ensure_proposal_id(md):
+    if 'proposal_id' not in md:
+        raise ValueError("You forgot the proposal id.")
+# RE.md_validator = ensure_proposal_id
+
+
+# be nice on segfaults
+import faulthandler
+faulthandler.enable()
+
+# set up logging framework
+import logging
+import sys
+
+handler = logging.StreamHandler(sys.stderr)
+fmt = logging.Formatter("%(asctime)-15s [%(name)5s:%(levelname)s] %(message)s")
+handler.setFormatter(fmt)
+handler.setLevel(logging.INFO)
+
+logging.getLogger('hxntools').addHandler(handler)
+logging.getLogger('hxnfly').addHandler(handler)
+logging.getLogger('ppmac').addHandler(handler)
+
+logging.getLogger('hxnfly').setLevel(logging.DEBUG)
+logging.getLogger('hxntools').setLevel(logging.DEBUG)
+logging.getLogger('ppmac').setLevel(logging.INFO)
+
+
+# Flyscan results are shown using pandas. Maximum rows/columns to use when
+# printing the table:
+pd.options.display.width = 180
+pd.options.display.max_rows = None
+pd.options.display.max_columns = 10
+
+# enable < shortcut to replace RE(
+from bluesky.utils import register_transform
+register_transform('RE', prefix='<')
 
 
 def remove_names_maybe(obj, names):
