--- profile_collection/startup/01-bluesky.py
@@ -1,36 +1,0 @@
-import functools
-from ophyd.commands import (setup_ophyd, wh_pos as _wh_pos)
-setup_ophyd()
-
-from ophyd.commands import *
-from bluesky.callbacks import *
-from bluesky.plans import *
-# from bluesky.spec_api import *
-from bluesky.utils import (install_qt_kicker, register_transform)
-from bluesky.global_state import (get_gs, abort, stop, resume)
-
-
-# The following line allows bluesky and pyqt4 GUIs to play nicely together:
-install_qt_kicker()
-
-gs = get_gs()
-RE = gs.RE
-RE.md['group'] = ''
-RE.md['config'] = {}
-RE.md['beamline_id'] = 'HXN'
-
-RE.verbose = True
-RE.ignore_callback_exceptions = False
-
-# Allow scans to be run by using the prefix '<' instead of typing RE(...)
-register_transform('RE', prefix='<')
-
-# The default is 'count_time', HXN has always used 'exposure_time':
-gs.MD_TIME_KEY = 'exposure_time'
-
-
-@functools.wraps(_wh_pos)
-def wh_pos(*args, **kwargs):
-    from datetime import datetime
-    print(datetime.now())
-    return _wh_pos(*args, **kwargs)
